requirements.txt
# Core
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Data & export
pandas>=2.0.0

# Optional: visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Dev / tests
pytest>=7.0.0
pytest-cov>=4.0.0
